moved file upload to a separate component
